合并

+++ src/models/structure.js
@@ -90,6 +90,9 @@ class Structure {
     if (node !== this.root) {
       let parent = this.findParent(node_id);
       data.name = getUniqName(this.getNames(parent, node_id), data.name);
+      if (!_.some(node.members, m => m._id.equals(ObjectId(data.admin)))) {
+        node.members.push({_id: ObjectId(data.admin)});
+      }
     }
     _.extend(node, data);
     return node;
@@ -120,6 +123,9 @@ class Structure {
     let node = this.findNodeById(node_id);
     if (node) {
       node.admin = member_id ? ObjectId(member_id) : null;
+      if (!_.some(node.members, m => m._id.equals(ObjectId(member_id)))) {
+        node.members.push({_id: member_id});
+      }
       return true;
     }
     return false;
@@ -165,7 +171,9 @@ class Structure {
     if (!_.isArray(node.members)) {
       node.members = [];
     }
-    node.members.push(member);
+    member.forEach(item => {
+      node.members.push(item);
+    });
     return member;
   }
 
@@ -205,6 +213,9 @@ class Structure {
     if (!node) {
       return false;
     }
+    if (node_id.equals(this.root._id)) {
+      return true;
+    }
     this._deleteMember(node, member_id);
     return true;
   }
@@ -239,6 +250,7 @@ class Structure {
 
   deletePosition(position_id, node_id) {
     node_id = ObjectId(node_id);
+    position_id = ObjectId(position_id);
     let node = this.findNodeById(node_id);
     if (!_.isArray(node.positions)) {
       node.positions = [];
@@ -247,7 +259,15 @@ class Structure {
     let index = _.findIndex(node.positions, pos => pos._id.equals(position_id));
     if (index >= 0) {
       node.positions.splice(index, 1);
-      node.members = _.reject(node.members, m => m.position.equals(position_id));
+      node.members = _.map(node.members, m => {
+        if (m.position && m.position.equals(position_id)) {
+          return {
+            _id: m._id
+          };
+        } else {
+          return m;
+        }
+      });
       return true;
     } else {
       return false;
@@ -299,6 +319,17 @@ class Structure {
   findPositionInfo(position_id) {
     position_id = ObjectId(position_id);
     return this._findPositionInfo(this.root, position_id);
+  }
+
+  addPositionToMember(node_id, member_id, position) {
+    node_id = ObjectId(node_id);
+    member_id = ObjectId(member_id);
+    position = ObjectId(position);
+    let node = this.findNodeById(node_id);
+    if (!_.some(node.positions, item => item._id.equals(position))) {
+      return null;
+    }
+    return _.some(node.members, m => (m._id.equals(member_id))&&(m.position = position)) ? {_id: member_id, position} : null;
   }
 
   findMemberDepartments(member_id, node, path) {
